Factor out list utils (#3234)
