+++ .changeset/lucky-jokes-warn.md
@@ -1,0 +1,5 @@
+---
+'@keystonejs/keystone': patch
+---
+
+Factored out utility `List` functions into a separate module.
