+++ packages/keystone/lib/ListTypes/list.js
@@ -12,100 +12,25 @@ const {
   flatten,
   zipObj,
   createLazyDeferred,
-  upcase,
 } = require('@keystonejs/utils');
 const { parseListAccess } = require('@keystonejs/access-control');
 const { logger } = require('@keystonejs/logger');
-
-const graphqlLogger = logger('graphql');
-const keystoneLogger = logger('keystone');
-
+const {
+  preventInvalidUnderscorePrefix,
+  keyToLabel,
+  labelToPath,
+  labelToClass,
+  opToType,
+  mapNativeTypeToKeystoneType,
+  getDefautlLabelResolver,
+} = require('./utils');
 const {
   LimitsExceededError,
   ValidationFailureError,
   throwAccessDenied,
 } = require('./graphqlErrors');
 
-const preventInvalidUnderscorePrefix = str => str.replace(/^__/, '_');
-
-const keyToLabel = str => {
-  let label = str
-    .replace(/([a-z])([A-Z])/g, '$1 $2')
-    .split(/\s|_|\-/)
-    .filter(i => i)
-    .map(upcase)
-    .join(' ');
-
-  // Retain the leading underscore for auxiliary lists
-  if (str[0] === '_') {
-    label = `_${label}`;
-  }
-  return label;
-};
-
-const labelToPath = str =>
-  str
-    .split(' ')
-    .join('-')
-    .toLowerCase();
-
-const labelToClass = str => str.replace(/\s+/g, '');
-
-const opToType = {
-  read: 'query',
-  create: 'mutation',
-  update: 'mutation',
-  delete: 'mutation',
-};
-
-const mapNativeTypeToKeystoneType = (type, listKey, fieldPath) => {
-  const { Text, Checkbox, Float } = require('@keystonejs/fields');
-
-  const nativeTypeMap = new Map([
-    [
-      Boolean,
-      {
-        name: 'Boolean',
-        keystoneType: Checkbox,
-      },
-    ],
-    [
-      String,
-      {
-        name: 'String',
-        keystoneType: Text,
-      },
-    ],
-    [
-      Number,
-      {
-        name: 'Number',
-        keystoneType: Float,
-      },
-    ],
-  ]);
-
-  if (!nativeTypeMap.has(type)) {
-    return type;
-  }
-
-  const { name, keystoneType } = nativeTypeMap.get(type);
-
-  keystoneLogger.warn(
-    { nativeType: type, keystoneType, listKey, fieldPath },
-    `Mapped field ${listKey}.${fieldPath} from native JavaScript type '${name}', to '${keystoneType.type.type}' from the @keystonejs/fields package.`
-  );
-
-  return keystoneType;
-};
-
-const getDefautlLabelResolver = labelField => item => {
-  const value = item[labelField || 'name'];
-  if (typeof value === 'number') {
-    return value.toString();
-  }
-  return value || item.id;
-};
+const graphqlLogger = logger('graphql');
 
 module.exports = class List {
   constructor(
